requirements/base.txt
pytz==2022.1  # https://github.com/stub42/pytz
python-slugify==6.1.2  # https://github.com/un33k/python-slugify
Pillow==9.1.1  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.2.0  # https://github.com/evansd/whitenoise
redis==4.3.3  # https://github.com/redis/redis-py
hiredis==2.0.0  # https://github.com/redis/hiredis-py

# Django
# ------------------------------------------------------------------------------
django==3.2.14  # pyup: < 4.0  # https://www.djangoproject.com/
django-environ==0.9.0  # https://github.com/joke2k/django-environ
django-model-utils==4.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.51.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.14.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.6  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.2.0  # https://github.com/jazzband/django-redis


django-htmlmin==0.11.0 # https://django-tailwind.readthedocs.io/en/latest/installation.html

python-webpack-boilerplate==0.0.10
readtime==1.1.1
django-admin-honeypot==1.1.0 # https://pypi.org/project/django-admin-honeypot/
django-jet2==1.0.12
feedparser==6.0.8
google-api-python-client==1.4.1

django-stdimage==5.3.0 # http://johannes.maron.family/django-stdimage/
django-countries-plus==2.0.0
django-parler==2.3
django-rosetta==0.9.8
loguru==0.5.3

django-autoslug==1.9.8
django-crontab==0.7.1
django-htmx==1.9.0
schedule==1.1.0

schedule==1.1.0
gevent==21.12.0

crispy-tailwind==0.5.0
django-mathfilters==1.0.0

django-tinymce==3.4.0
django-filebrowser-no-grappelli==4.0.0 #https://github.com/smacker/django-filebrowser-no-grappelli
django-google-translate==1.1  #https://pypi.org/project/django-google-translate/
